Removed unnecessary size_t casts. Also added resource for white noise generator

+++ Source/DSP/TextureLayer.cpp
@@ -29,20 +29,21 @@ void TextureLayer::process(juce::AudioBuffer<float>& buffer, const size_t numSam
 
 std::vector<float> TextureLayer::_generateNoiseSamples(const size_t numSamples) const
 {
-    // Random number generator
+    // Random number generator. Based on: https://github.com/hollance/synth-recipes/blob/main/recipes/white-noise.markdown
+
     const long randomSeed = random();
     std::linear_congruential_engine<std::uint_fast32_t, 196314165, 907633515, 0> rng{ static_cast<std::uint_fast32_t>(randomSeed) };
 
     // For scaling generated numbers to the range [-1, 1]
     constexpr float scale = (2.0f / static_cast<float>(rng.max() + 1.0f));
 
-    std::vector<float> buffer(static_cast<size_t>(numSamples));
+    std::vector<float> buffer(numSamples);
 
     for (size_t i = 0; i < numSamples; ++i)
     {
         const std::uint_fast32_t randomInt = rng();
         const float randomFloat = (static_cast<float>(randomInt) * scale) - 1.0f;
-        buffer[static_cast<size_t>(i)] = randomFloat;
+        buffer[i] = randomFloat;
     }
 
     return buffer;
